requirements.txt
scikit-learn
pandas
python-dotenv
nltk
torch
transformers
datasets
yacs
matplotlib
tqdm